requirements.txt
fastapi 
uvicorn 
sqlalchemy 
pymysql 
passlib 
python-jose 
python-multipart
aiocron
pydantic[email]
aiosmtplib